backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anthropic==0.13.0
anyio==3.7.1
arxiv==2.2.0
async-timeout==5.0.1
asyncpg==0.29.0
attrs==25.3.0
backoff==2.2.1
beautifulsoup4==4.13.5
blockbuster==1.5.25
boto3==1.40.27
botocore==1.40.27
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
cloudpickle==3.1.1
cryptography==44.0.3
curl_cffi==0.13.0
dataclasses-json==0.6.7
deepdiff==8.6.1
distro==1.9.0
docstring_parser==0.17.0
duckduckgo_search==4.5.0
emoji==2.14.1
fastapi==0.115.6
feedparser==6.0.11
filelock==3.19.1
filetype==1.2.0
fireworks-ai==0.11.2
forbiddenfruit==0.1.4
frozenlist==1.7.0
fsspec==2025.9.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-aiplatform==1.91.0
google-cloud-bigquery==3.30.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
groq==0.31.1
grpc-google-iam-v1==0.14.2
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
hf-xet==1.1.9
httpcore==1.0.9
httpx==0.25.2
httpx-sse==0.4.1
huggingface-hub==0.34.4
idna==3.10
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema_rs==0.29.1
kay==0.1.2
langchain==0.3.27
langchain-community==0.3.29
langchain-core>=0.3.75,<2.0.0
langchain-google-vertexai>=0.1.0
langchain-groq==0.3.8
langchain-openai==0.0.8
langchain-robocorp==0.0.4
langchain-text-splitters==0.3.11
langdetect==1.0.9
langgraph==0.6.7
langgraph-api==0.4.15
langgraph-checkpoint==2.1.1
langgraph-checkpoint-sqlite==2.0.11
langgraph-cli==0.4.2
langgraph-prebuilt==0.6.4
langgraph-runtime-inmem==0.12.0
langgraph-sdk==0.2.6
langgraph-studio==0.0.5
langserve==0.0.32
langsmith==0.4.27
lxml==6.0.1
marshmallow==3.26.1
multidict==6.6.4
mypy_extensions==1.1.0
nltk==3.9.1
numpy==1.26.4
openai==1.107.0
orderly-set==5.5.0
orjson==3.11.3
ormsgpack==1.10.0
packaging==23.2
pdfminer.six==20231228
pgvector==0.2.5
pillow==11.3.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyJWT==2.10.1
pypdf==6.0.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.6
PyYAML==6.0.2
RapidFuzz==3.14.1
regex==2025.9.1
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
s3transfer==0.14.0
sgmllib3k==1.0.0
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.8
SQLAlchemy==2.0.43
sqlite-vec==0.1.6
sse-starlette==2.1.3
starlette==0.41.3
structlog==25.4.0
tabulate==0.9.0
tenacity==8.5.0
tiktoken==0.5.2
tokenizers==0.22.0
tomli_w==1.2.0
tqdm==4.67.1
truststore==0.10.4
types-protobuf==4.25.0.20240417
types-requests==2.32.4.20250809
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
unstructured==0.12.5
unstructured-client==0.22.0
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
wikipedia==1.4.0
wrapt==1.17.3
xmltodict==0.13.0
xxhash==3.5.0
yarl==1.20.1
zstandard==0.24.0
